Updated version with dedicated messages,
