+++ src/css/main.css
@@ -6410,3 +6410,6 @@ p, button {
 
 .text-center1 {
   opacity: 0; }
+
+.three-buttons {
+  padding-top: 3rem; }
