+++ src/sass/main.scss
@@ -116,3 +116,7 @@ p, button {
 .text-center1 {
   opacity: 0;
 }
+
+.three-buttons {
+  padding-top: 3rem;
+}
